[8.19][ska] relocate functional_enterprise_search
