+++ .buildkite/ftr_search_stateful_configs.yml
@@ -1,12 +1,12 @@
 disabled:
   # Base config files, only necessary to inform config finding script
-  - x-pack/test/functional_enterprise_search/base_config.ts
+  - x-pack/solutions/search/test/functional_enterprise_search/base_config.ts
   - x-pack/solutions/search/test/api_integration/config.ts
 
   # Cypress configs, for now these are still run manually
-  - x-pack/test/functional_enterprise_search/cypress.config.ts
-  - x-pack/test/functional_enterprise_search/visual_config.ts
-  - x-pack/test/functional_enterprise_search/cli_config.ts
+  - x-pack/solutions/search/test/functional_enterprise_search/cypress.config.ts
+  - x-pack/solutions/search/test/functional_enterprise_search/visual_config.ts
+  - x-pack/solutions/search/test/functional_enterprise_search/cli_config.ts
 
 defaultQueue: 'n2-4-spot'
 enabled:
