+++ x-pack/solutions/search/test/functional_enterprise_search/apps/enterprise_search/with_host_configured/index.ts
