+++ x-pack/solutions/search/test/functional_enterprise_search/ftr_provider_context.d.ts
