+++ x-pack/solutions/search/test/functional_enterprise_search/page_objects/workplace_search.ts
