+++ x-pack/solutions/search/test/functional_enterprise_search/services/app_search_client.ts
